pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "longevity-intelligence-platform"
version = "0.1.0"
description = "Personal health AI: biological age, mortality risk, and digital twin simulator"
requires-python = ">=3.11"
dependencies = [
    # Data
    "pandas>=2.2",
    "pyarrow>=15",
    "numpy>=1.26",
    "scipy>=1.13",
    "requests>=2.31",
    "tqdm>=4.66",

    # ML / Tree models
    "scikit-learn>=1.4",
    "lightgbm>=4.3",
    "xgboost>=2.0",
    "optuna>=3.6",
    "joblib>=1.3",

    # Survival analysis
    "lifelines>=0.27",
    "xgbse>=0.3",

    # Deep learning
    "torch>=2.3",
    "torchvision>=0.18",
    "timm>=1.0",

    # Causal inference
    "dowhy>=0.11",
    "econml>=0.15",

    # Explainability
    "shap>=0.45",
    "dice-ml>=0.11",

    # NLP
    "transformers>=4.40",

    # API + validation
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "python-multipart>=0.0.9",

    # AI coach
    "anthropic>=0.28",

    # Config + logging
    "pyyaml>=6.0",
    "structlog>=24.1",

    # Experiment tracking
    "mlflow>=2.13",

    # Database (conversation history)
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
train-bioage = "longevity.models.bioage.trainer:main"
train-mortality = "longevity.models.mortality.trainer:main"
download-nhanes = "longevity.data.nhanes.downloader:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src/longevity --cov-report=term-missing --cov-fail-under=80"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
